ep22 - shwoop up

+++ TrickingLibrary.Api/BackgroundServices/VideoEditing/VideoManager.cs
@@ -19,23 +19,28 @@ namespace TrickingLibrary.Api.BackgroundServices.VideoEditing
             _env = env;
         }
 
-        public string WorkingDirectory => _env.WebRootPath;
+        private string WorkingDirectory => _env.WebRootPath;
+
+        public string FFMPEGPath => Path.Combine(_env.ContentRootPath, "ffmpeg", "ffmpeg.exe");
 
         public bool Temporary(string fileName)
         {
             return fileName.StartsWith(TempPrefix);
         }
 
-        public bool TemporaryVideoExists(string fileName)
+        public bool TemporaryFileExists(string fileName)
         {
             var path = TemporarySavePath(fileName);
             return File.Exists(path);
         }
 
-        public void DeleteTemporaryVideo(string fileName)
+        public void DeleteTemporaryFile(string fileName)
         {
             var path = TemporarySavePath(fileName);
-            File.Delete(path);
+            if (File.Exists(path))
+            {
+                File.Delete(path);
+            }
         }
 
         public string DevVideoPath(string fileName)
@@ -44,7 +49,7 @@ namespace TrickingLibrary.Api.BackgroundServices.VideoEditing
         }
 
         public string GenerateConvertedFileName() => $"{ConvertedPrefix}{DateTime.Now.Ticks}.mp4";
-        public string GenerateThumbnailFileName() => $"{ThumbnailPrefix}{DateTime.Now.Ticks}.png";
+        public string GenerateThumbnailFileName() => $"{ThumbnailPrefix}{DateTime.Now.Ticks}.jpg";
 
         public async Task<string> SaveTemporaryVideo(IFormFile video)
         {
